Changed BodyHealthChangeEvent#setNewHealth from int to double

+++ src/main/java/bodyhealth/api/events/BodyPartHealthChangeEvent.java
@@ -61,7 +61,7 @@ public class BodyPartHealthChangeEvent extends Event implements Cancellable {
         return newHealth;
     }
 
-    public void setNewHealth(int health) {
+    public void setNewHealth(double health) {
         newHealth = health;
     }
